Prepare TPC for multi-step
