+++ assets/src/domain/eventEditor/ui/tickets/ticketPriceCalculator/hooks/index.ts
@@ -1,3 +1,5 @@
+export { default as useMutatePrices } from './useMutatePrices';
+
 export { default as usePriceModifier } from './usePriceModifier';
 
 export { default as useOnSubmitPrices } from './useOnSubmitPrices';
