+++ assets/src/domain/eventEditor/ui/tickets/ticketPriceCalculator/utils/constants.ts
@@ -1,4 +1,4 @@
 import { Ticket } from '@edtrServices/apollo/types';
 
 // 'name' is only required for modal title
-export const ticketFieldsToUse: Array<keyof Partial<Ticket>> = ['id', 'name', 'price', 'reverseCalculate'];
+export const ticketFieldsToUse: Array<keyof Partial<Ticket>> = ['id', 'isTaxable', 'name', 'price', 'reverseCalculate'];
